TSP portion of GA is now working

+++ SA.cpp
@@ -170,7 +170,7 @@ int SA::solverSA(std::vector<std::vector<float>> &adjacancyMatrix, std::vector<I
 				tmpItems.pop_back();
 			}
 
-			int newProfit = calculateProfit(adjacancyMatrix, valuableItemsMatrix, calcPath, tmpItems, 
+			float newProfit = calculateProfit(adjacancyMatrix, valuableItemsMatrix, calcPath, tmpItems, 
 											noOfCities, noOfItems, knapsack);
 
 			if (newProfit > currProfit || randFraction() < exp((currProfit - newProfit) / T))
